chore(): various minor tweaks (#45)
* refactor(assets): moved fonts to assets folder, updated imports
* not defaulting to vtree mode for now
* defaulting to turned on grid for now
* moved modules around / renamed / restructured ui
* split huge 'main' view into smaller functions
* styling adjustments
* refactor(main): renamed entry point main to index

* docs(README): updated instructions about geometry caching

* chore(pkg): added url of git repo

* chore(deps): switched to using @jscad/io instead of individual serializers

* chore(pkg): added automatic generation of all binaries after version bump
