+++ packages/jscad-desktop/assets/fonts/Open_Sans/OpenSans-BoldItalic.ttf
