+++ packages/jscad-desktop/assets/fonts/Open_Sans/OpenSans-Italic.ttf
