+++ packages/jscad-desktop/index.html
@@ -10,20 +10,20 @@
             font-family: 'Open Sans';
             font-style: normal;
             font-weight: 400;
-            src: local('OpenSans'), local('OpenSans'), url(fonts/Open_Sans/OpenSans-Regular.ttf) format('truetype');
+            src: local('OpenSans'), local('OpenSans'), url(assets/fonts/Open_Sans/OpenSans-Regular.ttf) format('truetype');
           }
 
           @font-face {
             font-family: 'Open Sans';
             font-style: italic;
             font-weight: 400;
-            src: local('OpenSans'), local('OpenSans'), url(fonts/Open_Sans/OpenSans-Italic.ttf) format('truetype');
+            src: local('OpenSans'), local('OpenSans'), url(assets/fonts/Open_Sans/OpenSans-Italic.ttf) format('truetype');
           }
 
           @font-face {
             font-family: 'Open Sans';
             font-style: bold;
-            src: local('OpenSans'), local('OpenSans'), url(fonts/Open_Sans/OpenSans-Bold.ttf) format('truetype');
+            src: local('OpenSans'), local('OpenSans'), url(assets/fonts/Open_Sans/OpenSans-Bold.ttf) format('truetype');
           }
 
           html{
@@ -97,8 +97,6 @@
           #busy{
             position: absolute;
             color: black;
-            top: 50px;
-            z-index: 100;
           }
           #status{
             position: absolute;
@@ -217,10 +215,6 @@
           }
           #appUpdates {
             visibility: visible;
-            position: absolute;
-            z-index: 99;
-            top: 40px;
-            right: 100px;
           }
           #appUpdates a{
             text-decoration:none;
